fix: upgrade to esm
