+++ jest.config.js
@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   restoreMocks: true,
   clearMocks: true,
   // collectCoverage: true,
@@ -14,5 +14,8 @@ module.exports = {
       statements: 100
     }
   },
-  testRegex: /\.test\.jsx?$/.source
+  testRegex: /\.test\.jsx?$/.source,
+  transform: {
+    '\\.[jt]sx?$': 'babel-jest'
+  }
 };
