+++ lib/index.cjs
@@ -1,11 +1,14 @@
-import GithubSlugger from 'github-slugger';
+'use strict';
+
+var GithubSlugger = require('github-slugger');
+
 let slugger;
 
-export function reset() {
+function reset() {
   slugger = new GithubSlugger();
 }
 
-export function gfmHeadingId({ prefix = '' } = {}) {
+function gfmHeadingId({ prefix = '' } = {}) {
   reset();
 
   return {
@@ -16,3 +19,6 @@ export function gfmHeadingId({ prefix = '' } = {}) {
     }
   };
 }
+
+exports.gfmHeadingId = gfmHeadingId;
+exports.reset = reset;
