+++ package.json
@@ -2,7 +2,20 @@
   "name": "marked-gfm-heading-id",
   "version": "1.0.0",
   "description": "marked GFM heading ids",
-  "main": "src/index.js",
+  "main": "./lib/index.cjs",
+  "module": "./src/index.js",
+  "browser": "./lib/index.umd.js",
+  "type": "module",
+  "files": [
+    "lib/",
+    "src/"
+  ],
+  "exports": {
+    ".": {
+      "import": "./src/index.js",
+      "require": "./lib/index.cjs"
+    }
+  },
   "keywords": [
     "marked",
     "extension"
@@ -11,7 +24,7 @@
     "test": "jest --verbose",
     "test:cover": "jest --coverage",
     "lint": "eslint .",
-    "build": "rollup -c rollup.config.esm.js"
+    "build": "rollup -c rollup.config.cjs"
   },
   "repository": {
     "type": "git",
@@ -30,6 +43,8 @@
     "github-slugger": "^1.4.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.19.6",
+    "@babel/preset-env": "^7.19.4",
     "@rollup/plugin-commonjs": "^23.0.2",
     "@rollup/plugin-node-resolve": "^15.0.1",
     "@semantic-release/changelog": "^6.0.1",
@@ -38,6 +53,7 @@
     "@semantic-release/github": "^8.0.6",
     "@semantic-release/npm": "^9.0.1",
     "@semantic-release/release-notes-generator": "^10.0.3",
+    "babel-jest": "^29.2.1",
     "eslint": "^8.26.0",
     "eslint-config-standard": "^17.0.0",
     "eslint-plugin-import": "^2.26.0",
