+++ rollup.config.cjs
@@ -1,0 +1,17 @@
+module.exports = [
+  {
+    input: 'src/index.js',
+    output: {
+      name: 'markedBidi',
+      file: 'lib/index.umd.js',
+      format: 'umd'
+    }
+  },
+  {
+    input: 'src/index.js',
+    output: {
+      file: 'lib/index.cjs',
+      format: 'cjs'
+    }
+  }
+];
